backend/requirements.txt
# Core Web Framework & Server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Settings & Data Validation
pydantic>=2.6.4
pydantic-settings>=2.0.0 # Use pydantic-settings for .env loading
python-dotenv>=1.0.0

# HTTP Requests & APIs
httpx>=0.27.0
google-generativeai>=0.5.0 # Direct Gemini (optional fallback)
openrouter-python>=1.0.0 # Verify exact package name via pip install openrouter-python or similar
python-telegram-bot[ext]>=21.0.1 # Include [ext] for Application builder if needed later
twilio>=8.0.0 # Verify latest stable version
deepgram-sdk>=3.1.4 # Verify latest stable version
elevenlabs>=1.0.0 # Verify latest stable version
requests>=2.31.0 # For simpler sync requests if needed

# Web Scraping & Parsing
beautifulsoup4>=4.12.0
lxml>=5.0.0
# selenium>=4.19.0 # Uncomment ONLY if dynamic scraping is implemented and Chrome installed in Dockerfile
# webdriver-manager>=4.0.0 # Uncomment ONLY if using Selenium

# PDF Generation
reportlab>=4.0.0

# Caching (In-Memory)
cachetools>=5.3.0

# Scheduling Background Tasks
schedule>=1.2.0

# Utilities
pytz # Often needed for timezone handling

# Add any other specific libraries identified during service implementation
# Example: aiosmtplib if switching to async email sending later